% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aphiaIDs.R
\name{get_aphiaid_from_lookup}
\alias{get_aphiaid_from_lookup}
\title{Consult a lookup table for the aphiaID.}
\usage{
get_aphiaid_from_lookup(sciname, lookup)
}
\arguments{
\item{sciname}{A Scientific name as a string.}

\item{lookup}{The named list containing key-value pairs of scientific names and aphiaIDs.}
}
\value{
Returns the appropriate aphiaID corresponding to the sciname.
}
\description{
This is the helper function that we use in the sapply when mutating the WORMS_species_aphia_id into existence.
}
