% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aphiaIDs.R
\name{get_unique_aphiaids}
\alias{get_unique_aphiaids}
\title{Get AphiaIDs for scientific names}
\usage{
get_unique_aphiaids(scinames)
}
\arguments{
\item{scinames}{A vector (dataframe column) containing the list of scientific names from a detection extract dataframe in
Surimi.}
}
\value{
Returns a named list with the scientific name as the key and the aphiaID as the value.
}
\description{
Takes a column of scientific names and creates a lookup table (read: named list) of the unique scientific names
against their aphia IDs. We can use worrms to query the WORMS REST service for the aphiaIDs, but doing it for every row is
time intensive in a way we don't want. This way, we can create the lookup client-side and then do all the querying only as
we need to.
}
