% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatos_workbook_to_ato.R
\name{glatos_workbook_to_ato}
\alias{glatos_workbook_to_ato}
\title{Convert GLATOS workbook to an ATO object.}
\usage{
glatos_workbook_to_ato(glatos_workbook)
}
\arguments{
\item{glatos_workbook}{Path to the glatos workbook.}
}
\value{
Returns an ATO object.
}
\description{
Takes a GLATOS workbook and returns an ATO object.
}
