% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_otn_file.R
\name{map_otn_file}
\alias{map_otn_file}
\title{Determine whether an input file is CSV or Parquet and pipe it into the correct mapping function.}
\usage{
map_otn_file(filename, derive = TRUE, coll_code = NULL)
}
\arguments{
\item{filename}{The path to the file to be processed.}

\item{derive}{Passed through to the mapping functions; determines whether or not receiver and tag metadata will be derived from the detection extract or not.}

\item{coll_code}{Passed through to the mapping functions; allows user to supply collectionCode if they are passing their own rcvr/tag metadata, which won't contain the collectionCode.}
}
\value{
The output of the appropriate mapping function.
}
\description{
Determine whether an input file is CSV or Parquet and pipe it into the correct mapping function.
}
