% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otn_to_ato.R
\name{otn_to_ato}
\alias{otn_to_ato}
\title{Convert OTN detection data to an ATO object.}
\usage{
otn_to_ato(otn_detections, otn_receivers = "", otn_tags = "")
}
\arguments{
\item{otn_detections}{The dataframe containing detection information.}

\item{otn_receivers}{The dataframe containing receiver information.}

\item{otn_tags}{The dataframe containing tag information.}
}
\value{
Returns an ATO object.
}
\description{
Takes an OTN detection extract and optionally receiver/tag metadata and returns an ATO object.
}
