% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{rollup}
\alias{rollup}
\title{Join output from Remora back onto its OTN detection extract.}
\usage{
rollup(detection_extract, remora_output)
}
\arguments{
\item{detection_extract}{Path to an OTN detection extract corresponding to the remora output in the second parameter.}

\item{remora_output}{Path to Remora's QC output corresponding to the OTN detection extract in the first parameter.}
}
\value{
The OTN detection extract, but with the remora QC attached as appropriate.
}
\description{
Take two parameters- an OTN detection extract and the output created by Remora on parsing that detection extract- and merge them back together such that the Remora QC columns are appended to the OTN extract,
preserving appropriate ordering and getting back all the OTN data. This function exists because, to get OTN data into Remora, we have to cut it up until it looks like IMOS data (this problem was the genesis of Surimi, in fact).
But that means the output from Remora has all IMOS-formatted columns and is missing some information, because we either had to discard it to get into IMOS format or because we can't re-synthesize it from what's in the IMOS
files. However, we do have enough information to join the two tables, thereby obviating the data loss problem by taking us all the way back to the original data, with a little something extra attached.
}
