% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatos_to_ato.R
\name{glatos_to_ato}
\alias{glatos_to_ato}
\title{Convert GLATOS detection data to an ATO object.}
\usage{
glatos_to_ato(glatos_detections, glatos_receivers = "")
}
\arguments{
\item{otn_detections}{The dataframe containing detection information.}

\item{otn_receivers}{The dataframe containing receiver information.}

\item{otn_tags}{The dataframe containing tag information.}
}
\value{
Returns an ATO object.
}
\description{
Takes a GLATOS detection sheet and optionally receiver metadata and returns an ATO object.
}
